refactor(atomix): remove atomix-core module
It contained only test classes. These have been moved (and in one case merged) into atomix-cluster
