+++ atomix/cluster/src/test/java/io/atomix/cluster/AtomixClusterTest.java
@@ -16,20 +16,90 @@
  */
 package io.atomix.cluster;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 import io.atomix.cluster.discovery.BootstrapDiscoveryProvider;
+import io.atomix.raft.partition.RaftPartitionGroupConfig;
+import io.atomix.raft.partition.RaftStorageConfig;
 import io.atomix.utils.net.Address;
+import java.io.File;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.junit.Rule;
 import org.junit.Test;
 
 /** Atomix cluster test. */
 public class AtomixClusterTest {
+  private static final int TIMEOUT_IN_S = 90;
+
+  @Rule public final AtomixClusterRule atomixClusterRule = new AtomixClusterRule();
+
+  @Test
+  public void testStopStartConsensus() throws Exception {
+    // given
+    final var atomix =
+        atomixClusterRule
+            .startAtomix(
+                1,
+                Arrays.asList(1),
+                (builder) -> {
+                  final var groupConfig =
+                      new RaftPartitionGroupConfig()
+                          .setName("raft")
+                          .setReplicationFactor(3)
+                          .setPartitionCount(7)
+                          .setMembers(Set.of("1"))
+                          .setStorageConfig(
+                              new RaftStorageConfig()
+                                  .setDirectory(
+                                      new File(
+                                              atomixClusterRule.getDataDir(),
+                                              "start-stop-consensus")
+                                          .getPath())
+                                  .setPersistedSnapshotStoreFactory(
+                                      new NoopSnapshotStoreFactory()));
+
+                  return builder.build();
+                })
+            .get(TIMEOUT_IN_S, TimeUnit.SECONDS);
+
+    // when
+    final var stopFuture = atomix.stop();
+
+    // then
+    assertThat(stopFuture).succeedsWithin(TIMEOUT_IN_S, TimeUnit.SECONDS);
+    assertThat(stopFuture).isDone();
+  }
+
+  @Test
+  public void shouldFailStartAfterStop() throws Exception {
+    // given
+    final var atomix =
+        atomixClusterRule
+            .startAtomix(1, Arrays.asList(1), AtomixClusterBuilder::build)
+            .get(TIMEOUT_IN_S, TimeUnit.SECONDS);
+    atomix.stop().get(TIMEOUT_IN_S, TimeUnit.SECONDS);
+
+    // when
+    try {
+      atomix.start().get(TIMEOUT_IN_S, TimeUnit.SECONDS);
+      fail("Expected ExecutionException");
+    } catch (final ExecutionException ex) {
+      // then
+      assertTrue(ex.getCause() instanceof IllegalStateException);
+      assertEquals("AtomixCluster instance shutdown", ex.getCause().getMessage());
+    }
+  }
 
   @Test
   public void testBootstrap() throws Exception {
