+++ atomix/cluster/src/test/java/io/atomix/cluster/NoopSnapshotStore.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core;
+package io.atomix.cluster;
 
 import io.camunda.zeebe.snapshots.PersistedSnapshot;
 import io.camunda.zeebe.snapshots.PersistedSnapshotListener;
