+++ atomix/cluster/src/test/java/io/atomix/cluster/messaging/TestUnicastService.java
@@ -14,12 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.test.messaging;
+package io.atomix.cluster.messaging;
 
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import io.atomix.cluster.messaging.ManagedUnicastService;
-import io.atomix.cluster.messaging.UnicastService;
 import io.atomix.utils.net.Address;
 import java.util.Map;
 import java.util.Set;
