--- atomix/core/src/test/resources/logback.xml
@@ -1,33 +1,0 @@
-<!--
-  ~ Copyright 2017-present Open Networking Laboratory
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<configuration>
-  <appender class="ch.qos.logback.core.ConsoleAppender" name="STDOUT">
-    <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n
-      </pattern>
-    </encoder>
-  </appender>
-
-  <logger level="${root.logging.level:-INFO}" name="io.atomix.protocols.backup"/>
-  <logger level="${root.logging.level:-INFO}" name="io.atomix.raft"/>
-  <logger level="${root.logging.level:-INFO}" name="io.atomix.protocols.log"/>
-  <logger level="INFO" name="io.atomix.cluster.protocol"/>
-  <logger level="INFO" name="io.atomix.messaging"/>
-
-  <root level="${root.logging.level:-INFO}">
-    <appender-ref ref="STDOUT"/>
-  </root>
-</configuration>
